+++ DIMS-Core/DIMS-Core.DataAccessLayer/Repositories/Base/Repository.cs
@@ -6,7 +6,7 @@ using DIMS_Core.DataAccessLayer.Interfaces;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 
-namespace DIMS_Core.DataAccessLayer.Repositories
+namespace DIMS_Core.DataAccessLayer.Repositories.Base
 {
     /// <summary>
     /// Base Repository
