+++ DIMS-Core/DIMS-Core.DataAccessLayer/Repositories/TaskStateRepository.cs
@@ -7,6 +7,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using DIMS_Core.DataAccessLayer.Repositories.Base;
 using Task = System.Threading.Tasks.Task;
 
 namespace DIMS_Core.DataAccessLayer.Repositories
