+++ DIMS-Core/DIMS-Core.DataAccessLayer/Repositories/UserTaskRepository.cs
@@ -1,10 +1,6 @@
 using DIMS_Core.DataAccessLayer.Models;
 using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using DIMS_Core.DataAccessLayer.Repositories.Base;
 
 namespace DIMS_Core.DataAccessLayer.Repositories
 {
