+++ DIMS-Core/DIMS-Core.DataAccessLayer/Repositories/VUserTaskRepository.cs
@@ -1,0 +1,14 @@
+using DIMS_Core.DataAccessLayer.Models;
+using DIMS_Core.DataAccessLayer.Repositories.Base;
+
+namespace DIMS_Core.DataAccessLayer.Repositories
+{
+    public class VUserTaskRepository : ReadOnlyRepository<VUserTask>
+    {
+
+        public VUserTaskRepository(DimsCoreContext context) : base(context)
+        {
+
+        }
+    }
+}
